refactor: Switch 컴포넌트 외부 상태 기반으로 리팩토링 (#33)
