+++ src/shared/ui/atoms/switch/index.tsx
@@ -1,41 +1,26 @@
-import { useEffect, useState } from "react";
+import { useEffect, useRef } from "react";
 import { Animated, Easing, Pressable, View, ViewProps } from "react-native";
-import shadowStyles from "@shared/tokens/elevation";
-import { Thumb, Track } from "./variants";
+import { elevation } from "@shared/tokens";
+import { Thumb, Track } from "./switch-styles";
 
 interface Props extends Omit<ViewProps, "children"> {
-  value?: boolean;
-  defaultValue?: boolean;
-  onToggle?: (value: boolean) => void;
+  value: boolean; // 외부에서 전달되는 On/Off 상태
+  onToggle: () => void; // 토글 시 외부에 알리는 콜백
 }
 
-// 외부에서 값이 안 들어오면 기본으로 꺼진 상태 (defaultValue)
-export default function Switch({ value, defaultValue = false, onToggle, ...rest }: Props) {
-  // 외부 value가 없으면 defaultValue 사용
-  const [isOn, setIsOn] = useState(value ?? defaultValue);
+export default function Switch({ value, onToggle, ...rest }: Props) {
+  // value가 true면 1, false면 0
+  const animatedValue = useRef(new Animated.Value(value ? 1 : 0)).current;
 
-  useEffect(() => {
-    if (value !== undefined) setIsOn(value);
-  }, [value]);
-
-  // 스위치 토글 함수
-  const toggle = () => {
-    const newValue = !isOn;
-    setIsOn(newValue);
-    onToggle?.(newValue);
-  };
-
-  // 0: off, 1: on
-  const [animatedValue] = useState(new Animated.Value(isOn ? 1 : 0));
-
+  // value 상태에 따른 애니메이션
   useEffect(() => {
     Animated.timing(animatedValue, {
-      toValue: isOn ? 1 : 0,
+      toValue: value ? 1 : 0,
       duration: 200,
       easing: Easing.linear,
       useNativeDriver: false,
     }).start();
-  }, [isOn, animatedValue]);
+  }, [value, animatedValue]);
 
   const translateX = animatedValue.interpolate({
     inputRange: [0, 1],
@@ -43,10 +28,11 @@ export default function Switch({ value, defaultValue = false, onToggle, ...rest 
   });
 
   return (
-    <Pressable onPress={toggle} {...rest}>
+    // Press 시 onToggle 호출
+    <Pressable onPress={onToggle} {...rest}>
       {({ pressed }) => (
-        <View className={Track(isOn, pressed)} style={pressed ? shadowStyles.active : undefined}>
-          <Animated.View className={Thumb(isOn, pressed)} style={{ transform: [{ translateX }] }} />
+        <View className={Track(value, pressed)} style={pressed && elevation.active}>
+          <Animated.View className={Thumb(value, pressed)} style={{ transform: [{ translateX }] }} />
         </View>
       )}
     </Pressable>
