+++ src/shared/ui/atoms/switch/switch-styles.ts
